fix: debt-sources — приводить rental_id/scooter_id к int (raw SQL отдаёт строкой)
Без ::int rentalId приходил строкой: ломалось исключение собственной аренды
(currentChainIds.includes мимо), а scooterId-строка не находила имя скутера —
плашка показывала «#3» вместо модели.

+++ apps/api/src/routes/clients.ts
@@ -245,11 +245,11 @@ export async function clientsRoutes(app: FastifyInstance) {
       const raw = await db.execute(sql`
         SELECT
           dr.id AS "drId",
-          dr.rental_id AS "rentalId",
+          dr.rental_id::int AS "rentalId",
           dr.total::int AS "total",
           dr.deposit_covered::int AS "depositCovered",
           dr.note AS "note",
-          r.scooter_id AS "scooterId",
+          r.scooter_id::int AS "scooterId",
           r.status AS "status",
           r.archived_at AS "archivedAt",
           r.start_at AS "startAt",
